not the prettiest, but updating test cases
