+++ tests/test_mouselab.py
@@ -117,59 +117,3 @@ def test_nonsymmetric(setting, expected):
                 symmetric = False
 
     assert symmetric == expected
-
-click_orders = [[(1,2,3),(2,3,1),(3,1,2)],[(2,3),(3,2)],[(1,2),(2,1)],[(3,1),(1,3)]]
-extended_belief_state_test_case = [
-    {
-        "env"      : {
-            "name"             : "small_test_case",
-            "branching"        : [1, 2],
-            "reward_inputs"    : "depth",
-            "reward_dictionary": {1: Categorical([-500]), 2: Categorical([-60, 60])},
-        },
-        "structure": {
-            "layout" : {"0": [0, 0], "1": [0, -1], "2": [1, -1], "3": [-1, -1]},
-            "initial": "0",
-            "graph"  : {
-                "0": {"up": [0, "1"]},
-                "1": {"right": [0, "2"], "left": [0, "3"]},
-                "2": {},
-                "3": {},
-            },
-        },
-        "click_orders" : click_order
-    } for click_order in click_orders
-]
-
-
-@pytest.fixture(params=extended_belief_state_test_case)
-def setup_extended_belief_state(request):
-    """
-    Would be better to refactor this since test_exact.py has similar function
-    """
-    register(**request.param["env"])
-
-    env_params = {"cost":distance_graph_cost(max_penalty=None, distance_multiplier=1),
-                      "include_last_action":True,
-                      "mdp_graph_properties":get_structure_properties(request.param["structure"])}
-
-    yield MouselabEnv.new_symmetric_registered(request.param["env"]["name"],
-                                               **env_params), request.param["click_orders"]
-
-
-def test_extended_belief_state_hash(setup_extended_belief_state):
-    """
-    Test to see if hash includes last action for extended belief state
-    """
-    env, click_orders = setup_extended_belief_state
-
-    final_hashes = []
-    for click_order in click_orders:
-        env.reset()
-        for action in click_order:
-            env.step(action)
-        final_hashes.append(hash(env._state))
-
-    # all should be unique
-    assert len(set(final_hashes))==len(final_hashes)
-
